Fix but in list
when a changed function is set, deleting the first element (0-index) lead to a index out of range error.
Now the index will only be decreased if the last item was removed. That changes the behaviour of the selection slightly:
It will stay at the same place and only move up when the last item was removed.