Update README.md to include NatsDash application - Basically a NATS GUI
Super happy to include NatsDash that I have been recently working on. The initial version of it used Terminal.Gui using c#, I almost gave up due to how complicated and verbose it was. Then I switched to Tview and it is such a joy to work with. Thanks for all the work, couldn't have been possible without tview.